Fix compilation errors for Waveshare Touch LCD 2 build
Configs: add MARAUDER_WAVESHARE_TOUCH_LCD_2 to MENU DEFINITIONS
section in src/configs.h.  Earlier I copied the display-defines block
but missed that KEY_X/Y/W/H, ICON_W/H, BANNER_TIME, BUTTON_PADDING
live in a SEPARATE per-board block at the bottom of configs.h.
Values match V8 / CYD_2USB (same 240-wide screen geometry).

platformio.ini: add mikalhart/MicroNMEA to lib_deps.  Marauder
#include <MicroNMEA.h> unconditionally in src/GpsInterface.h, even
though HAS_GPS is off on this board, so we still need to link it.

src/WiFiScan.cpp: guard WIFI_AUTH_WPA3_ENTERPRISE with
'defined(WIFI_AUTH_WPA3_ENTERPRISE)' inside the existing HAS_IDF_3
block.  The enum is newer than Arduino-ESP32 3.20017; HAS_IDF_3 is
defined but the enum itself is not.

src/WiFiScan.cpp: in writeSidecar(), guard the gps_obj.getDatetime()
call with HAS_GPS.  gps_obj is only instantiated when HAS_GPS is on,
and we deliberately left it off (no GPS pinout mapped yet).

src/MenuFunctions.cpp: cast both 'min(...)' calls to unsigned int.
BUTTON_SCREEN_LIMIT is signed int but menu->list->size() returns
size_t (unsigned int), so std::min can't deduce a single _Tp.
Explicit (unsigned int) casts fix the overload resolution.

+++ src/MenuFunctions.cpp
@@ -4244,7 +4244,7 @@ void MenuFunctions::buildButtons(Menu *menu, int starting_index, const char* but
 
   this->menu_start_index = starting_index;
 
-  uint8_t visible_buttons = min(BUTTON_SCREEN_LIMIT, menu->list->size() - starting_index);
+  uint8_t visible_buttons = min((unsigned int)BUTTON_SCREEN_LIMIT, (unsigned int)(menu->list->size() - starting_index));
 
   for (uint8_t i = 0; i < visible_buttons; i++) {
     MenuNode node = menu->list->get(starting_index + i);
@@ -4315,7 +4315,7 @@ void MenuFunctions::displayCurrentMenu(int start_index)
       display_obj.tft.setTextSize(1);
     #endif
 
-    for (uint16_t i = start_index; i < min(start_index + BUTTON_SCREEN_LIMIT, current_menu->list->size()); i++)
+    for (uint16_t i = start_index; i < min((unsigned int)(start_index + BUTTON_SCREEN_LIMIT), (unsigned int)current_menu->list->size()); i++)
     {
       if (!current_menu || !current_menu->list || i >= current_menu->list->size())
         continue;
